Function trillium_async_std::run

pub fn run(handler: impl Handler)

Runs a trillium handler in a sync context with default config

Runs a trillium handler on the async-std runtime with default configuration. See crate::config for what the defaults are and how to override them

This function will block the current thread until the server shuts down